Use one manager for all integration test cases
